chore(battery): rename app.c to battery.c
